// erzahler/src/models/enumeration/order-display-enum.ts
export enum OrderDisplay {
  AIRLIFT = 'Airlift',
  CONVOY = 'Convoy',
  DISBAND = 'Disband',
  HOLD = 'Hold',
  INVALID = 'Invalid (Holding)',
  MOVE = 'Move',
  MOVE_CONVOYED = 'Move Convoyed',
  NUKE = 'Nuke',
  OFFER_TECH = 'Offer Nuke Tech',
  RECEIVE_TECH = 'Receive Nuke Tech From',
  RUSH_BUILD = 'Rush Build',
  SUPPORT = 'Support',
  SUPPORT_CONVOYED = 'Support Convoyed',
  TRANSFER = 'Transfer To'
}

export enum OrderSetType {
  ORDERS = 'Orders',
  PREDICTION = 'Prediction',
  SUGGESTION = 'Suggestion'
}